requirements.txt
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
ffmpeg-python==0.2.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
packaging==24.1
pycparser==2.22
PySoundFile==0.9.0.post1
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.4
setuptools==74.0.0
srt==3.5.3
sympy==1.13.2
torch==2.4.0
torchaudio==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
